requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.23.0
scipy>=1.9.0
pyyaml>=6.0
matplotlib>=3.6.0
pytest>=7.0.0
pydantic>=1.10.0
